Split audio code into its own module
