+++ index.html
@@ -6,6 +6,8 @@
 </head>
 <body>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r128/three.min.js"></script>
-    <script src="assets/js/main.js"></script>
+    <script src="assets/js/game.js"></script>
+    <script src="assets/js/audio.js"></script>
+    <script src="assets/js/ui.js"></script>
 </body>
 </html>
